+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -5,6 +5,12 @@
 * Drop support for .NET 6 as this target is no longer supported and add .NET 8 target.
   ([#2160](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2160))
 
+* Add `OpenTelemetry.Instrumentation.StackExchangeRedis` meter. Supported metrics:
+  * `redis.client.request.duration`,
+  * `redis.client.queue.duration`,
+  * `redis.client.network.duration`.
+  ([#1982](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1982)) 
+
 ## 1.9.0-beta.1
 
 Released 2024-Jul-23